pyproject.toml
[tool.poetry]
name = "cookbook-importer"
version = "0.1.0"
description = ""
authors = ["Edward Q. Bridges <ebridges@roja.cc>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
"yaml-1.3" = "^0.1.0"
PyYAML = "^5.3.1"
"ruamel.yaml" = "^0.16.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
